minitl: fixes to new format library
Various MSVC specific fixes

+++ src/motor/core/src/memory/streams.cc
@@ -6,7 +6,7 @@
 
 namespace Motor {
 
-MemoryStream::MemoryStream(minitl::Allocator& allocator, u64 size)
+MemoryStream::MemoryStream(minitl::Allocator& allocator, u32 size)
     : m_memory(allocator, size)
     , m_size(size)
     , m_capacity(0)
@@ -17,7 +17,7 @@ MemoryStream::~MemoryStream()
 {
 }
 
-void MemoryStream::resize(u64 size)
+void MemoryStream::resize(u32 size)
 {
     if(!m_memory.resize(size))
     {
@@ -25,14 +25,14 @@ void MemoryStream::resize(u64 size)
     }
 }
 
-void MemoryStream::write(const void* buffer, u64 size_)
+void MemoryStream::write(const void* buffer, u32 size_)
 {
     if(size_ + m_size > m_capacity) resize(m_size + size_);
     memcpy(m_memory + m_size, buffer, motor_checked_numcast< size_t >(size_));
     m_size = m_size + size_;
 }
 
-void MemoryStream::erase(u64 count)
+void MemoryStream::erase(u32 count)
 {
     motor_assert_format(m_size >= count, "erasing {0} bytes from a stream of {1} bytes", count,
                         m_size);
